% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dataset_url.R
\name{create_dataset_url}
\alias{create_dataset_url}
\title{Create CSIRO dataset URL}
\usage{
create_dataset_url(variable, model, rcp, year_range)
}
\description{
Parameters used in determining the dataset URL
}
