% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heat_vulnerability_index.R
\name{get_heat_vulnerability_index}
\alias{get_heat_vulnerability_index}
\title{NSW Heat Vulnerability Index to ABS Statistical Area Level 1 2016}
\usage{
get_heat_vulnerability_index(sa1)
}
\arguments{
\item{sa1}{text, SA1 2016 code}
}
\description{
The NSW Heat Vulnerability Index (HVI) dataset identifies areas to monitor where
populations in the Sydney Greater Metropolitan Area are more vulnerable to the
adverse effects of urban heat, as of Summer 2015-2016. HVI utilises indicators
for exposure, sensitivity and adaptive capacity to calculate an overall heat
vulnerability index. Expressed through the data, a vulnerability of 1 represents a
combination of low exposure, low sensitivity and/or high adaptive capacity.
A vulnerability of 5 represents high exposure, high sensitivity and/or low adaptive capacity.
The calculation of HVI and the inputs to the exposure, sensitivity and adaptive
capacity indicators are explained in the metadata. The HVI data is aggregated to
the Australian Bureau of Statistics (ABS) Statistical Area Level 1 (SA1) polygon
dataset to enable spatial analysis to support local policy and decision making.
It can be used in conjunction with the NSW urban vegetation cover dataset for
the same time period for broader analysis of the relationship of heat to green cover.
}
\details{
https://datasets.seed.nsw.gov.au/dataset/nsw-heat-vulnerability-index-to-abs-statistical-area-level-1-2016
}
