% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_urban_heat_island_value.R
\name{get_urban_heat_island_value}
\alias{get_urban_heat_island_value}
\title{Get Urban Heat Island}
\usage{
get_urban_heat_island_value(mb)
}
\arguments{
\item{mb}{text, Mesh block 2016 code}
}
\description{
The Urban Heat Island (UHI) dataset measures the effects of urbanisation on land
surface temperatures across Sydney Greater Metropolitan Area for the Summer of 2015-2016.
UHI shows the variation of temperature to a non-urban vegetated reference, such as
heavily wooded areas or national parks around Sydney. Derived from the analysis of
thermal and infrared data from Landsat satellite, the dataset has been combined
with the Australian Bureau of Statistics (ABS) Mesh Block polygon dataset to
provide a mean UHI temperature that enables multi-scale spatial analysis of
the relationship of heat to green cover.
}
\details{
https://datasets.seed.nsw.gov.au/dataset/nsw-urban-heat-island-to-modified-mesh-block-2016
}
