% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_sa1.R
\name{join_sa1}
\alias{join_sa1}
\title{Join Statistical Area 1 2016 to Mesh Block 2016}
\usage{
join_sa1(x)
}
\arguments{
\item{x}{data.frame, data with MB_2016_CODE}
}
\description{
MB 2016 is present in the GANF, we attach SA1 using a concordance dataset \code{mb2016_to_sa12016}
}
