% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_heat_vulnerability_index.R
\name{map_heat_vulnerability_index}
\alias{map_heat_vulnerability_index}
\title{Map heat vulnerability index data to ordinal names}
\usage{
map_heat_vulnerability_index(x)
}
\arguments{
\item{x}{list, result from \code{get_heat_vulnerability_index}}
}
\description{
1 represents low exposure, low sensitivity or high adaptive capacity and 5
represents high exposure, high sensitivity or low adaptive capacity.
}
