% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_urban_vegetation_cover_all.R
\name{map_urban_vegetation_cover_all}
\alias{map_urban_vegetation_cover_all}
\title{Wrangle urban vegetation cover results}
\usage{
map_urban_vegetation_cover_all(x)
}
\arguments{
\item{x, }{list, that from \code{map_urban_vegetation_cover_all}}
}
\value{
tibble / data.frame
}
\description{
Wrangle list into  data.frame
}
