% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mb2016_to_sa12016.R
\docType{data}
\name{mb2016_to_sa12016}
\alias{mb2016_to_sa12016}
\title{Meshblock 2016 to SA1 2016 concorddance}
\format{
A data frame with xx rows and yy variables:
\describe{
\item{MB_CODE_2016 }{chr, meshblock 2016 code}
\item{SA1_MAINCODE_2016}{chr, SA1 2016 code}
}
}
\source{
\url{https://www.abs.gov.au}
}
\usage{
mb2016_to_sa12016
}
\description{
Mesh Blocks are the smallest geographical area defined by the Australian Bureau
of Statistics (ABS) and form the building blocks for the larger regions of
the Australian Statistical Geography Standard (ASGS). All other statistical
areas or regions are built up from or, approximated by whole Mesh Blocks.
They broadly identify land use such as residential, commercial, primary production
and parks, etc.
}
\details{
This dataset was originally found on https://www.abs.gov.au

\url{https://www.abs.gov.au/ausstats/abs@.nsf/Lookup/by\%20Subject/1270.0.55.001~July\%202016~Main\%20Features~Mesh\%20Blocks\%20(MB)~10012}
}
\keyword{datasets}
