% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_api.R
\name{query_api}
\alias{query_api}
\title{Query ArcGIS REST Service using httr2}
\usage{
query_api(url, where_query_string)
}
\arguments{
\item{url}{string, API URL}

\item{where_query_string}{string, where clause e.g. "SA1_MAIN16=123456"}
}
\description{
Hopefully they're all the same! Heat Vulnerability Index
}
