% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_address_coords.R
\name{return_address_coords}
\alias{return_address_coords}
\title{Query the flat minified GNAF dataset}
\usage{
return_address_coords(street_number, street_name, locality, postcode)
}
\arguments{
\item{street_number}{street number without any unit/apartment info}

\item{street_name}{street name}

\item{locality}{locality/suburb}

\item{postcode}{postcode. Must be 4 digits}
}
\description{
A filter dplyr::filter function on the columns in \code{gnaf}. Missing values
will return a T in the filter step, essentially meaning 'everything' for that
column
}
