% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnaf-leaflet-example.R
\name{run_gnaf_leaflet_shiny}
\alias{run_gnaf_leaflet_shiny}
\title{A shiny app with GNAF address selection and view on leaflet}
\usage{
run_gnaf_leaflet_shiny()
}
\description{
Start a shiny app with address selectors that show results on leaflet
}
